requirements.txt
Django==1.6.4
argparse==1.2.1
distribute==0.6.24
django-autoslug==1.7.2
django-autoslug-field==0.2.3
django-debug-toolbar==1.2
django-mptt==0.6.0
django-mptt-admin==0.1.8
eav-django==1.4.5
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
